+++ week01/day5/employee_db.py
@@ -1,4 +1,5 @@
 import json
+import utils
 
 from pathlib import Path
 
@@ -11,6 +12,10 @@ def load_employees():
             return json.load(file)
 
     except FileNotFoundError:
+        return [] 
+    
+    except json.JSONDecodeError:
+        print("Error: emp.json contains invalid JSON.")
         return []
     
 def save_employees(employees):
@@ -28,20 +33,20 @@ def add_employee():
 
     emp = {
 
-        "id": int(input("ID : ")),
+        "id": utils.get_int("ID : "),
 
         "name": input("Name : "),
 
         "department": input("Department : "),
 
-        "salary": float(input("Salary : "))
+        "salary": utils.get_float("Salary : ")
     }
     for item in employees:
         if item["id"] == emp["id"]:
             print("Employee ID already exists")
             return   
-        else:
-            employees.append(emp)
+        
+    employees.append(emp)
 
     save_employees(employees)
 
@@ -84,6 +89,63 @@ def search_employee(employee_id: int):
         
     print("Employee not found")
            
+def update_employee(employee_id: int) -> None:
+    employees = load_employees()
+
+    for emp in employees:
+
+        if emp["id"] == employee_id:
+
+            print("Employee Found")
+
+            # Ask for new values here
+            
+            print("Current Name: ", emp["name"])
+            new_name = input("(press Enter to keep current):")
+            if new_name:
+                    emp["name"] = new_name
+                    
+                    
+            print("Current Department: ", emp["department"])
+            new_dept = input("(press Enter to keep current):")
+            if new_dept:
+                    emp["department"] = new_dept
+                    
+            print("Current Salary: ", emp["salary"])
+            new_sal = input("(press Enter to keep current):")
+            if new_sal:
+                    emp["salary"] = utils.get_float(new_sal)
+
+           
+
+            # Save employees here
+            save_employees(employees)
+            print("Employee Updated.")
+            return
+
+    print("Employee not found.")
     
 
-    
+def delete_employee(employee_id: int) -> None:
+    employees = load_employees()
+
+    for emp in employees:
+
+        if emp["id"] == employee_id:
+
+            # Display employee
+
+            confirm = input("Delete employee? (Y/N): ")
+
+            if confirm.strip().lower() == "y":
+
+                # Remove employee
+                employees.remove(emp)
+                # Save employees
+                save_employees(employees)
+
+                print("Employee deleted successfully.")
+
+            return
+
+    print("Employee not found.")